Fixed the "ownership" of a couple of files.
Also fixed trac bug #1. The theoretical limit of entities in a directory is now
264672, that should be sufficient for a while and will increase dramatically if
BLOCK_SIZE is ever increased.

This was almost completely implemented already, all I had to do was make sure
the size field of a directory's inode would actually store the number of bytes
allocated to that directory, not the number of files in the directory, and it
pretty much worked. The other changes are just modifications of the various
pieces of code that list the contents of a directory.